Correct Rotation based on MRA expectation
The rotation flag in MRAs is based around the direction the mister has to rotate it, rather than the direction the video is rotated. This value used to not matter, but now the MiSTer sends the rotation information in the DV infoframe so that scalers can automatically rotate the image.